+++ db/migrate/20171124202705_add_parent_id_at_last_commit_to_file_items.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# Add parent ID at last commit to file items to track whether a file was moved
+class AddParentIdAtLastCommitToFileItems < ActiveRecord::Migration[5.1]
+  def change
+    add_column :file_items, :parent_id_at_last_commit, :bigint, null: true
+  end
+end